Separate user interface into new package Duke.Ui
Classes for the user interface (using JavaFX) are found in the same package as the core program logic.

The code becomes quite disorganised, with many class files in the same root folder.

Let's create a new package Duke.Ui to encapsulate all classes for the user interface.

This adds hierarchy and organisation to the project.
